Split out cjs and esm configs
